Move createIntl function to a separate file to support react server components
