+++ packages/react-intl/src/components/provider.tsx
@@ -4,29 +4,12 @@
  * See the accompanying LICENSE file for terms.
  */
 
-import {
-  CreateIntlFn,
-  FormatMessageFn,
-  IntlCache,
-  createIntl as coreCreateIntl,
-  formatMessage as coreFormatMessage,
-  createIntlCache,
-} from '@formatjs/intl'
+import {IntlCache, createIntlCache} from '@formatjs/intl'
 import * as React from 'react'
-import type {IntlConfig, IntlShape, ResolvedIntlConfig} from '../types'
-import {
-  DEFAULT_INTL_CONFIG,
-  assignUniqueKeysToParts,
-  invariantIntlContext,
-  shallowEqual,
-} from '../utils'
+import type {IntlConfig, IntlShape} from '../types'
+import {DEFAULT_INTL_CONFIG, invariantIntlContext, shallowEqual} from '../utils'
 import {Provider} from './injectIntl'
-
-import {
-  FormatXMLElementFn,
-  PrimitiveType,
-  isFormatXMLElementFn,
-} from 'intl-messageformat'
+import {createIntl} from './createIntl'
 
 interface State {
   /**
@@ -62,102 +45,6 @@ function processIntlConfig<P extends IntlConfig = IntlConfig>(
     wrapRichTextChunksInFragment: config.wrapRichTextChunksInFragment,
     defaultRichTextElements: config.defaultRichTextElements,
   }
-}
-
-function assignUniqueKeysToFormatXMLElementFnArgument<
-  T extends Record<
-    string,
-    | PrimitiveType
-    | React.ReactNode
-    | FormatXMLElementFn<React.ReactNode, React.ReactNode>
-  > = Record<
-    string,
-    | PrimitiveType
-    | React.ReactNode
-    | FormatXMLElementFn<React.ReactNode, React.ReactNode>
-  >,
->(values?: T): T | undefined {
-  if (!values) {
-    return values
-  }
-  return Object.keys(values).reduce((acc: T, k) => {
-    const v = values[k]
-    ;(acc as any)[k] = isFormatXMLElementFn<React.ReactNode>(v)
-      ? assignUniqueKeysToParts(v)
-      : v
-    return acc
-  }, {} as T)
-}
-
-const formatMessage: FormatMessageFn<React.ReactNode> = (
-  config,
-  formatters,
-  descriptor,
-  rawValues,
-  ...rest
-) => {
-  const values = assignUniqueKeysToFormatXMLElementFnArgument(rawValues)
-  const chunks = coreFormatMessage(
-    config,
-    formatters,
-    descriptor,
-    values as any,
-    ...rest
-  )
-  if (Array.isArray(chunks)) {
-    return React.Children.toArray(chunks)
-  }
-  return chunks as any
-}
-
-/**
- * Create intl object
- * @param config intl config
- * @param cache cache for formatter instances to prevent memory leak
- */
-export const createIntl: CreateIntlFn<
-  React.ReactNode,
-  IntlConfig,
-  IntlShape
-> = (
-  {defaultRichTextElements: rawDefaultRichTextElements, ...config},
-  cache
-) => {
-  const defaultRichTextElements = assignUniqueKeysToFormatXMLElementFnArgument(
-    rawDefaultRichTextElements
-  )
-  const coreIntl = coreCreateIntl<React.ReactNode>(
-    {
-      ...DEFAULT_INTL_CONFIG,
-      ...config,
-      defaultRichTextElements,
-    },
-    cache
-  )
-
-  const resolvedConfig: ResolvedIntlConfig = {
-    locale: coreIntl.locale,
-    timeZone: coreIntl.timeZone,
-    fallbackOnEmptyString: coreIntl.fallbackOnEmptyString,
-    formats: coreIntl.formats,
-    defaultLocale: coreIntl.defaultLocale,
-    defaultFormats: coreIntl.defaultFormats,
-    messages: coreIntl.messages,
-    onError: coreIntl.onError,
-    defaultRichTextElements,
-  }
-
-  return {
-    ...coreIntl,
-    formatMessage: formatMessage.bind(
-      null,
-      resolvedConfig,
-      // @ts-expect-error fix this
-      coreIntl.formatters
-    ),
-    // @ts-expect-error fix this
-    $t: formatMessage.bind(null, resolvedConfig, coreIntl.formatters),
-  } as any
 }
 
 export default class IntlProvider extends React.PureComponent<
