+++ packages/react-intl/tests/unit/components/plural.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import FormattedPlural from '../../../src/components/plural'
 import {mountFormattedComponentWithProvider} from '../testUtils'
-import {createIntl} from '../../../src/components/provider'
+import {createIntl} from '../../../src/components/createIntl'
 import {IntlShape} from '@formatjs/intl'
 import {render} from '@testing-library/react'
 import {LDMLPluralRule} from '@formatjs/ecma402-abstract'
